requirements.txt
# requirements.txt
# Core dependencies
pandas==2.0.3
numpy==1.24.3
python-dotenv==1.0.0

# OpenAI and tokenization
openai>=1.0.0
tiktoken>=0.5.0

# Streamlit (for future GUI)
streamlit==1.31.0

# Utilities
tqdm==4.66.1
click==8.1.7

# Data validation
pydantic==2.5.0

# Testing (optional)
pytest==7.4.3
pytest-cov==4.1.0

# Development tools (optional)
black==23.12.0
pylint==3.0.3
mypy==1.7.1